refactor(web): reorganize IDE layout, terminal, and shared code
- Extract useEditorGroups hook from ide-layout (615→289 lines)
- Extract useIdeKeyboardShortcuts hook
- Extract lib/ide/breadcrumb with matchNavItem, getBreadcrumbParts
- Split mock-terminal (725→269 lines): file-contents, commands, prompt
- Extract explorerTree to consts/explorer-tree.ts
- Add NavItem type to nav-items
- Remove dead apps/web compose-refs (unused, packages/ui has own)
- Breadcrumbs use shared getBreadcrumbParts
