+++ apps/web/src/components/ide/breadcrumbs.tsx
@@ -4,14 +4,8 @@ import { cn } from "@portfolio/ui";
 import { ChevronRight, Code, Copy, Eye } from "lucide-react";
 import { useTranslations } from "next-intl";
 import { navItems } from "@/consts/nav-items";
+import { getBreadcrumbParts } from "@/lib/ide/breadcrumb";
 import type { ViewMode } from "./view-mode";
-
-function matchNavItem(pathname: string) {
-  return navItems.find((item) => {
-    if (item.href === "/") return pathname === "/";
-    return pathname.startsWith(item.href);
-  });
-}
 
 interface BreadcrumbsProps {
   onCopy?: () => void;
@@ -27,20 +21,7 @@ export function Breadcrumbs({
   onCopy,
 }: BreadcrumbsProps) {
   const t = useTranslations("ide");
-  const parts = ["portfolio", "src"];
-  const navItem = matchNavItem(pathname);
-
-  if (navItem) {
-    if (pathname.startsWith("/blog") && pathname !== "/blog") {
-      parts.push("blog");
-      const slug = pathname.replace("/blog/", "");
-      parts.push(`${slug}.mdx`);
-    } else if (navItem.href === "/blog") {
-      parts.push("blog", "index.mdx");
-    } else {
-      parts.push(navItem.fileName);
-    }
-  }
+  const parts = getBreadcrumbParts(pathname, navItems);
 
   return (
     <div className="flex shrink-0 items-center justify-between gap-2 overflow-hidden border-border border-b bg-background px-2 py-1.5 text-[11px] text-muted-foreground sm:px-4 sm:py-1">
@@ -63,9 +44,9 @@ export function Breadcrumbs({
         })}
       </div>
       <div
+        aria-label="Editor actions"
         className="flex shrink-0 select-none items-center"
         role="group"
-        aria-label="Editor actions"
       >
         {onCopy && (
           <>
@@ -75,12 +56,16 @@ export function Breadcrumbs({
               title={t("copyContent")}
               type="button"
             >
-              <Copy className="size-3.5" aria-hidden />
+              <Copy aria-hidden className="size-3.5" />
             </button>
-            <span className="mx-0.5 h-3 w-px bg-border" aria-hidden />
+            <span aria-hidden className="mx-0.5 h-3 w-px bg-border" />
           </>
         )}
-        <div className="flex items-center rounded-sm" role="group" aria-label="View mode">
+        <div
+          aria-label="View mode"
+          className="flex items-center rounded-sm"
+          role="group"
+        >
           <button
             className={cn(
               "flex cursor-pointer items-center justify-center rounded p-1.5 transition-colors",
@@ -92,7 +77,7 @@ export function Breadcrumbs({
             title={t("preview")}
             type="button"
           >
-            <Eye className="size-3.5" aria-hidden />
+            <Eye aria-hidden className="size-3.5" />
           </button>
           <button
             className={cn(
@@ -105,7 +90,7 @@ export function Breadcrumbs({
             title={t("source")}
             type="button"
           >
-            <Code className="size-3.5" aria-hidden />
+            <Code aria-hidden className="size-3.5" />
           </button>
         </div>
       </div>
