+++ apps/web/src/components/ide/terminal/file-contents.ts
@@ -1,0 +1,147 @@
+/**
+ * Mock file contents for the terminal `cat` command.
+ * Paths resolve: about.md -> src/about.md
+ * Directories (src, apps, packages) cannot be used with cat
+ */
+export const FILE_CONTENTS: Record<string, string[]> = {
+  "tsconfig.json": [
+    "{",
+    '  "compilerOptions": {',
+    '    "target": "ES2022",',
+    '    "module": "ESNext",',
+    '    "moduleResolution": "bundler",',
+    '    "strict": true',
+    "  }",
+    "}",
+  ],
+  ".env": ["# Local env vars", "NODE_ENV=development"],
+  "package.json": [
+    "{",
+    '  "name": "portfolio",',
+    '  "version": "0.0.0",',
+    '  "private": true,',
+    '  "type": "module",',
+    '  "scripts": {',
+    '    "dev": "turbo dev",',
+    '    "build": "turbo build"',
+    "  }",
+    "}",
+  ],
+  "readme.md": [
+    "# Portfolio",
+    "",
+    "A developer portfolio built with Next.js.",
+    "",
+    "## Getting started",
+    "",
+    "```bash",
+    "bun install",
+    "bun run dev",
+    "```",
+  ],
+  "about.md": [
+    "# About",
+    "",
+    "> A bit about me",
+    "",
+    "I'm a software engineer with a passion for building backend",
+    "& web applications. Currently actively seeking new opportunities.",
+    "",
+    "---",
+    "",
+    "## Experience",
+    "",
+    "### Freelance — *2019 - Present*",
+    "Freelance Developer / Open Source Contributor",
+    "",
+    "### Tokyo School — *2024 - Present*",
+    "PCAP Python, Computer Programming",
+    "",
+    "### Insergal — *2018 - 2019*",
+    "Sales Assistant, Marketing",
+    "",
+    "### Insergal — *2018 - 2019*",
+    "Automotive Mechanic",
+  ],
+  "src/about.md": [
+    "# About",
+    "",
+    "> A bit about me",
+    "",
+    "I'm a software engineer with a passion for building backend",
+    "& web applications. Currently actively seeking new opportunities.",
+    "",
+    "---",
+    "",
+    "## Experience",
+    "",
+    "### Freelance — *2019 - Present*",
+    "Freelance Developer / Open Source Contributor",
+    "",
+    "### Tokyo School — *2024 - Present*",
+    "PCAP Python, Computer Programming",
+    "",
+    "### Insergal — *2018 - 2019*",
+    "Sales Assistant, Marketing",
+    "",
+    "### Insergal — *2018 - 2019*",
+    "Automotive Mechanic",
+  ],
+  "src/welcome.tsx": [
+    'import { Separator } from "@portfolio/ui";',
+    'import { FeaturedProjects } from "@/components/home/featured-projects";',
+    'import { SocialNav } from "@/components/home/social-nav";',
+    'import { getGitHubRepos } from "@/lib/github";',
+    "",
+    "export const metadata = {",
+    '  title: "Martin Vila",',
+    '  description: "developer · gamer · self-taught",',
+    "};",
+    "",
+    "const socials = [",
+    '  { name: "github", href: "https://github.com/kWAYTV", icon: Github },',
+    "  ...",
+    "];",
+    "",
+    "export default async function Home() {",
+    "  const repos = await getGitHubRepos();",
+    "  return (",
+    '    <main className="space-y-6">',
+    "      <h1>Martin Vila</h1>",
+    "      ...",
+    "    </main>",
+    "  );",
+    "}",
+  ],
+  "src/projects.ts": [
+    'import { Octokit } from "@octokit/rest";',
+    "",
+    "export async function getRepositories(): Promise<Repository[]> {",
+    "  const { data } = await octokit.repos.listForUser({",
+    '    username: "kWAYTV",',
+    '    sort: "updated",',
+    "    per_page: 100,",
+    "  });",
+    "  return data.map(({ name, description, ... }) => ({ name, description, ... }));",
+    "}",
+    "",
+    'export const sortOptions = ["stars", "updated", "created", "name"] as const;',
+  ],
+  "src/blog/index.mdx": [
+    "---",
+    "title: Blog",
+    "description: Quiet notes from current work",
+    "---",
+    "",
+    "# Blog",
+    "",
+    "Quiet notes from current work. More entries will fall in here",
+    "as they are published — kept chronologically, nothing fancy.",
+    "",
+    "## Latest Posts",
+    "",
+    "- **Shipping the Wrong Thing** — *Feb 23, 2026*",
+    "- **The Side Project Graveyard** — *Dec 22, 2025*",
+    "- **Hello World** — *Nov 29, 2025*",
+  ],
+};
